requirement.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.12.14
charset-normalizer==3.4.1
defusedxml==0.7.1
Django==4.2.18
django-filter==24.3
djangorestframework==3.15.2
httpie==3.2.4
idna==3.10
importlib-metadata==8.5.0
Markdown==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
psycopg2==2.9.10
pygments==2.19.1
PySocks==1.7.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
sqlparse==0.5.3
typing-extensions==4.12.2
urllib3==2.2.3
zipp==3.20.2